CRUD Country

+++ src/com/example/movie/dao/CountryDao.java
@@ -7,6 +7,7 @@ import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 
 import javax.transaction.Transactional;
+import java.util.List;
 
 public class CountryDao {
     private final SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
@@ -45,9 +46,27 @@ public class CountryDao {
     }
 
     @Transactional
-    public void delete(Country country) {
+    public void deleteById(int id) throws NotFound {
         Session session = sessionFactory.openSession();
+        session.beginTransaction();
+        Country country = session.get(Country.class, id);
+
+        if (country == null) {
+            session.close();
+            throw new NotFound();
+        }
+
         session.delete(country);
+        session.getTransaction().commit();
         session.close();
+    }
+
+    @Transactional
+    public List<Country> getAll() {
+        Session session = sessionFactory.openSession();
+        List<Country> countryList = session.createQuery("FROM Country ORDER BY id").list();
+        session.close();
+
+        return countryList;
     }
 }
